goose_farmer_server/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-anymail==12.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
idna==3.10
psycopg2==2.9.10
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
